Refactor: FeedController
