+++ src/main/java/world/trecord/controller/users/UserController.java
@@ -2,7 +2,6 @@ package world.trecord.controller.users;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -11,8 +10,6 @@ import world.trecord.config.security.CurrentUser;
 import world.trecord.controller.ApiResponse;
 import world.trecord.service.comment.CommentService;
 import world.trecord.service.comment.response.UserCommentsResponse;
-import world.trecord.service.feedcontributor.FeedContributorService;
-import world.trecord.service.feedcontributor.response.UserFeedContributorListResponse;
 import world.trecord.service.userrecordlike.UserRecordLikeService;
 import world.trecord.service.userrecordlike.response.UserRecordLikeListResponse;
 import world.trecord.service.users.UserContext;
@@ -26,7 +23,6 @@ import world.trecord.service.users.response.UserInfoResponse;
 public class UserController {
 
     private final UserService userService;
-    private final FeedContributorService feedContributorService;
     private final CommentService commentService;
     private final UserRecordLikeService userRecordLikeService;
 
@@ -61,11 +57,5 @@ public class UserController {
     public ApiResponse<UserRecordLikeListResponse> getUserRecordLikes(@PageableDefault(sort = "createdDateTime", direction = Sort.Direction.ASC) Pageable pageable,
                                                                       @CurrentUser UserContext userContext) {
         return ApiResponse.ok(userRecordLikeService.getUserRecordLikeList(userContext.getId()));
-    }
-
-    @GetMapping("/invited")
-    public ApiResponse<Page<UserFeedContributorListResponse>> getUserParticipatingFeeds(@PageableDefault(sort = "createdDateTime", direction = Sort.Direction.DESC) Pageable pageable,
-                                                                                        @CurrentUser UserContext userContext) {
-        return ApiResponse.ok(feedContributorService.getUserParticipatingFeeds(userContext.getId(), pageable));
     }
 }
